pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "liao"
dynamic = ["version"]
description = "Liao (聊) - Vision-based GUI interaction assistant with LLM integration"
readme = "README.md"
license = "GPL-3.0-or-later"
authors = [
    {name = "Liao Contributors"}
]
keywords = ["liao", "llm", "chat", "automation", "vision", "ocr", "desktop", "gui", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Desktop Environment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"

dependencies = [
    "PySide6>=6.4.0",
    "requests>=2.28.0",
    "pyautogui>=0.9.53",
    "pyperclip>=1.8.2",
    "Pillow>=9.3.0",
    "numpy>=1.24.0",
    "pywin32>=305",
]

[project.optional-dependencies]
ocr = [
    "easyocr>=1.6.2",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
all = [
    "easyocr>=1.6.2",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
dev = [
    "pytest>=7.2.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/cycleuser/Liao"
Documentation = "https://github.com/cycleuser/Liao#readme"
Repository = "https://github.com/cycleuser/Liao"
Issues = "https://github.com/cycleuser/Liao/issues"

[project.scripts]
liao = "liao.cli:main"

[project.gui-scripts]
liao-gui = "liao.cli:main"

[tool.setuptools.dynamic]
version = {attr = "liao.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
liao = ["gui/i18n/*.json"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
